requirements.txt
SQLAlchemy~=2.0.19
pydantic==2.1.1
fastapi~=0.100.0
selenium~=4.10.0
python-dotenv~=1.0.0
pydantic[email]